Call recording encoder/format choice 3/3
PS1: Adds a user option between WB AMR (the default) or HE-AAC for the encoding of call recording audio. Also changes the output format of the file appropriately (WB AMR now uses the WB_AMR output format and HE_AAC chooses the MPEG_4 output. File extensions are adjust accordingly as well).
This was in reponse to people complaining the quality of call recording audio is poor. It's noticibly better using HE-AAC, but filesizes are larger.

PS2: Removed Whitespace

Change-Id: Iffcdd3ba6f9aa80e238bbcc5db51809a4721711e
https://github.com/CM11MOD/android_frameworks_base/commit/e9ad694127ef139e170e52049f1b927d5d6927c2